% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_fun.R
\name{epi_xp_tp}
\alias{epi_xp_tp}
\alias{epi_dc3}
\alias{epi_dc6}
\alias{epi_dcmax}
\alias{epi_dcmed}
\alias{epi_ds3}
\alias{epi_ds6}
\alias{epi_dsmax}
\alias{epi_dsmed}
\alias{epi_prop}
\alias{epi_st}
\title{Compute epidemiological features from time series}
\usage{
epi_xp_tp(.data, y)

epi_dc3(.data, y)

epi_dc6(.data, y)

epi_dcmax(.data, y)

epi_dcmed(.data, y)

epi_ds3(.data, y)

epi_ds6(.data, y)

epi_dsmax(.data, y)

epi_dsmed(.data, y)

epi_prop(.data, y)

epi_st(.data, y)
}
\arguments{
\item{.data}{.data .data a \code{tsibble} object.}

\item{y}{y character. Reference variable with numeric values.}
}
\value{
a \code{tibble} object.
}
\description{
Following
}
\section{Functions}{
\itemize{
\item \code{epi_xp_tp()}: Compute Xp and Tp

\item \code{epi_dc3()}: Compute Dc3

\item \code{epi_dc6()}: Compute Dc6

\item \code{epi_dcmax()}: Compute Dcmax

\item \code{epi_dcmed()}: Compute Dcmed

\item \code{epi_ds3()}: Compute Ds3

\item \code{epi_ds6()}: Compute Ds6

\item \code{epi_dsmax()}: Compute Dsmax

\item \code{epi_dsmed()}: Compute Dsmed

\item \code{epi_prop()}: Compute P+ (Prop)

\item \code{epi_st()}: Compute ST

}}
