% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_length.R
\name{feat_length}
\alias{feat_length}
\title{Count the length of periods that meets criteria}
\usage{
feat_length(.data, y, a_op = "gte", a, b_op = "gte", b, isolated = FALSE)
}
\arguments{
\item{.data}{a \code{tsibble} object.}

\item{y}{character. Reference variable with numeric values.}

\item{a_op, b_op}{character. Operator, \code{gte} = greater than or equal, \code{lte} = less than or equal, \code{gt} = greater than, \code{lt} = less than, \code{e} = equal.}

\item{a}{integer. Length of period threshold.}

\item{b}{integer. Value threshold applied to \code{y}.}

\item{isolated}{logical. Consider only isolated events, ie. surrounded by zeros. On this case, \code{a} and \code{a_op} are not considered.}
}
\value{
a \code{tibble} object.
}
\description{
This function will count the length of periods that meets criteria for each key on a tsibble object.
}
\details{
Example: How many periods have at least 3 consecutive observations (\code{a_op = "gte", a = 3}) with values equal or greater than 5 (\code{b_op = "gte", b = 5})?
}
\examples{
# How many periods have at least 3 consecutive observations with value equal or greater than 5

example_data <- tsibble::tsibble(
cod = rep(1, 10),
time = 1:10,
value = c(8,15,20,0,0,0,0,5,9,12),
key = cod,
index = time
)

feat_length(.data = example_data, y = "value", a_op = "gte", a = 3, b_op = "gte", b = 5)

# How many periods have isolate observations with values equal or greater to 5

example_data <- tsibble::tsibble(
cod = rep(1, 11),
time = 1:11,
value = c(0,16,0,0,8,3,0,10,0,2,0), key = cod, index = time
)

feat_length(.data = example_data, y = "value",
a_op = "e", a = 1, b_op = "gte", b = 5, isolated = TRUE)

}
