% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_length.R
\name{feat_length_stat}
\alias{feat_length_stat}
\title{Compute a statistics on the length of periods that meets criteria}
\usage{
feat_length_stat(.data, y, b, b_op = "e", stat)
}
\arguments{
\item{.data}{a \code{tsibble} object.}

\item{y}{character. Reference variable with numeric values.}

\item{b}{b integer. Value threshold applied to \code{y}.}

\item{b_op}{character. Operator, \code{gte} = greater than or equal, \code{lte} = less than or equal, \code{gt} = greater than, \code{lt} = less than, \code{e} = equal.}

\item{stat}{character. A statistic to be calculated on \code{y}. One of: max, min, mean, median, sd, var.}
}
\value{
a \code{tibble} object.
}
\description{
This function will compute a statistics on the length of periods that meets criteria for each key on a tsibble object.
}
\details{
Example: What is the median length (\code{stat = "median"}) of consecutive observations with values greater or equal to 5 (\code{b = 5, b_op = "gte"})?
}
\examples{
# What is the median length of consecutive observations with values greater or equal to 5

example_data <- tsibble::tsibble(
cod = rep(1, 10),
time = 1:10,
value = c(8,15,20,0,0,0,0,5,9,12),
key = cod,
index = time
)

feat_length_stat(.data = example_data, y = "value",
b = 5, b_op = "gte", stat = "median")

}
