% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_max_min.R
\name{feat_max}
\alias{feat_max}
\alias{feat_min}
\title{Isolate rows with observed maximum or minimum value}
\usage{
feat_max(.data, y)

feat_min(.data, y)
}
\arguments{
\item{.data}{a \code{tsibble} object.}

\item{y}{character. Reference variable with numeric values.}
}
\value{
a \code{tsibble} object.
}
\description{
This function will filter -- for each key in a tsibble object -- the rows that contain maximum or minimum value.
}
\examples{
example_data <- tsibble::tsibble(
cod = rep(1, 10),
time = 1:10,
value = c(8,15,20,0,0,0,0,5,9,12),
key = cod,
index = time
)

feat_max(.data = example_data, y = "value")

}
