% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_prop.R
\name{feat_prop}
\alias{feat_prop}
\title{Compute the proportion of values that meets criteria}
\usage{
feat_prop(.data, y, b, b_op)
}
\arguments{
\item{.data}{a \code{tsibble} object.}

\item{y}{character. Reference variable with numeric values.}

\item{b}{integer. Value threshold applied to \code{y}.}

\item{b_op}{character. Operator, \code{gte} = greater than or equal, \code{lte} = less than or equal, \code{gt} = greater than, \code{lt} = less than, \code{e} = equal.}
}
\value{
a \code{tibble} object.
}
\description{
This function computes -- for each key on a tsibble object -- the proportion (from 0 to 1) of values that meets criteria.
}
\details{
Example: What is the proportion of values greater or equal to 5 (\code{b = 5, b_op = "gte"})?
}
\examples{

# What is the proportion of values greater or equal to 5?

example_data <- tsibble::tsibble(
cod = rep(1, 10),
time = 1:10,
value = c(8,15,20,0,0,0,0,5,9,12),
key = cod,
index = time
)

feat_prop(.data = example_data, y = "value", b = 5, b_op = "gte")


}
