% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_st.R
\name{feat_st}
\alias{feat_st}
\title{Compute the ratio between average trend and average seasonality}
\usage{
feat_st(.data, y, type)
}
\arguments{
\item{.data}{.data a \code{tsibble} object.}

\item{y}{character. Reference variable with numeric values.}

\item{type}{character. \code{mult} for multiplicative time series decomposition or \code{add} to additive time series composition.}
}
\value{
a \code{tibble} object.
}
\description{
This function will compute -- for each key on a tsibble object -- the ratio between average trend and average seasonality, obtained from a multiplicative or additive decomposition of the time series.
}
