% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trle.R
\name{trle}
\alias{trle}
\title{Run Length Encoding and return result as a tibble}
\usage{
trle(.data, y)
}
\arguments{
\item{.data}{a \code{tibble} object.}

\item{y}{character. A variable available on \code{.data}}
}
\value{
a \code{tibble} object.
}
\description{
Given a tibble object and a variable \code{y}, this function will count the number of occurrence of each element in \code{y} in the sequence that they appear, and return this count as a tibble object.
}
\examples{
example_1 <- tibble::tibble(
cod = rep(1, 10),
time = 1:10,
value = c(8,15,20,0,0,0,0,5,9,12)
)

trle(.data = example_1, y = "value")
}
\seealso{
\code{\link[=rle]{rle()}}
}
