% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_cut.R
\name{year_cut}
\alias{year_cut}
\title{Compute year given a time reference to cut}
\usage{
year_cut(data, cut, cut_type)
}
\arguments{
\item{data}{a \code{tsibble} object.}

\item{cut}{numeric. A month or a epidemiological week.}

\item{cut_type}{character. \code{month} or \code{epiweek}}
}
\value{
a \code{tsibble} object with \code{year_cut} variable.
}
\description{
This function creates a new variable \code{year_cut} by computing if the year of a date if bellow or above a \code{cut} reference.
}
\details{
A possible use case is to create a reference where the year  starts on April or on the 34 epidemiological week.
}
\examples{
example_data <- tsibble::tsibble(
cod = rep(1, 3653),
time = seq.Date(as.Date("2010-01-01"), as.Date("2020-01-01"), by = "day"),
count = runif(3653),
key = cod,
index = time
)

example_data \%>\%
year_cut(cut = 35, cut_type = "epiweek")

}
