% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_length.R
\name{feat_length}
\alias{feat_length}
\title{Compute length of periods}
\usage{
feat_length(.data, y, a_op = "gte", a, b_op = "gte", b, isolated = FALSE)
}
\arguments{
\item{.data}{a \code{tsibble} object.}

\item{y}{character. Reference variable with numeric values.}

\item{a_op, b_op}{character. Operator, \code{gte} = greater than or equal, \code{lte} = less than or equal, \code{gt} = greater than, \code{lt} = less than, \code{e} = equal.}

\item{a}{integer. Length of period threshold.}

\item{b}{integer. Value threshold applied to \code{y}.}

\item{isolated}{logical. Consider only isolated events, ie. surrounded by zeros. On this case, \code{a} and \code{a_op} are not considered.}
}
\value{
A tibble.
}
\description{
Compute length of periods
}
\examples{
# How many at least 3 consecutive periods with value equal or greater to 5

example_1 <- tsibble::tsibble(
cod = rep(1, 10),
time = 1:10,
value = c(8,15,20,0,0,0,0,5,9,12),
key = cod,
index = time
)

feat_length(.data = example_1, y = "value", a_op = "gte", a = 3, b_op = "gte", b = 5)

# How many isolate periods with value equal or greater to 5

example_4 <- tsibble::tsibble(
cod = rep(1, 11),
time = 1:11,
value = c(0,16,0,0,8,3,0,10,0,2,0),key = cod, index = time
)

feat_length(.data = example_4, y = "value", a_op = "e", a = 1, b_op = "gte", b = 5, isolated = TRUE)

}
