% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trle.R
\name{trle}
\alias{trle}
\title{Run Length Encoding with tibble result}
\usage{
trle(.data, y)
}
\arguments{
\item{.data}{a \code{tibble} object.}

\item{y}{character. A variable available on \code{.data}}
}
\value{
a \code{tibble} object.
}
\description{
Run Length Encoding with tibble result
}
\examples{
example_1 <- tibble::tibble(
cod = rep(1, 10),
time = 1:10,
value = c(8,15,20,0,0,0,0,5,9,12)
)

trle(.data = example_1, y = "value")
}
